requirements.txt
aws-cdk-lib==2.200.1
constructs>=10.0.0,<11.0.0
discord.py==2.5.2
python-dotenv==1.0.1
requests==2.32.4
pyyaml==6.0.2
flask==3.0.3
mangum==0.19.0
asgiref==3.8.1
discord-interactions==0.4.0